+++ src/main/kotlin/online/bingzi/luck/perms/bridge/spring/boot/starter/event/priority/EventPriority.kt
@@ -1,0 +1,12 @@
+package online.bingzi.luck.perms.bridge.spring.boot.starter.event.priority
+
+/**
+ * 事件优先级枚举
+ */
+enum class EventPriority {
+    LOWEST,
+    LOW,
+    NORMAL,
+    HIGH,
+    HIGHEST
+}